ENH: Treat sinks in ENHD more robustly [skip ci]
``enhd_flowlines_nx`` function now replaces all 0 values in the ``tocomid`` column of ENHD
with the negative of their corresponding ``comid`` values. This ensures
all sinks are unique and treated accordingly for topological sorting
and other network analysis. The difference are in the returned
``label2comid`` dictionary and ``onnetwork_sorted`` which will contain
the negative values for the sinks.